+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - declare compatibility with `python3.11`
+- allow parametrization of `spi_max_speed_hz` during `C1101` class instantiation
+  to solve [issue-128] 
 
 ### Changed
 - `CC1101.transmit`: raise `RuntimeError` instead of `Exception` when
@@ -133,3 +135,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.2.0]: https://github.com/fphammerle/python-cc1101/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/fphammerle/python-cc1101/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/fphammerle/python-cc1101/releases/tag/v1.0.0
+[issue-128]: https://github.com/fphammerle/python-cc1101/issues/128